Attempt at fixing some bugs in game.ex.
Updated variables to always refer to current player instead of player or updated player, for consistency and easier time to understand what the code is doing since the current player is the one being affected in most cases.
Fixed issue with owner and state not being stored in variables before trying to be accessed.
defp for handle_call at the bottom of the file changed to def and added the @impl tag, defp doesn't work for an implementation of a callback if previously defined using def.
TODO: Fix return values in the cond statement for handle_normal_roll().
TODO: Implement the is_property() method for the property module or fix the check for the handle_call of :buy_property.